yolov7train/requirements.txt
imutils==0.5.4
numpy==1.17.4
opencv_python_headless==4.6.0.66
tensorflow==2.10.0
asposestorage==1.0.2
Pillow==9.3.0
torch==1.13.0
tqdm==4.64.0
asposestorage==1.0.2
